move stuff but also fix chat stuff
